+++ database/sekai/bond.go
@@ -17,13 +17,13 @@ type Bond struct {
 	// ID of the ent.
 	ID int `json:"id,omitempty"`
 	// GameID holds the value of the "game_id" field.
-	GameID int `json:"game_id,omitempty"`
+	GameID int64 `json:"game_id,omitempty"`
 	// GroupID holds the value of the "group_id" field.
-	GroupID int `json:"group_id,omitempty"`
+	GroupID int64 `json:"group_id,omitempty"`
 	// CharacterId1 holds the value of the "character_id1" field.
-	CharacterId1 int `json:"character_id1,omitempty"`
+	CharacterId1 int64 `json:"character_id1,omitempty"`
 	// CharacterId2 holds the value of the "character_id2" field.
-	CharacterId2 int `json:"character_id2,omitempty"`
+	CharacterId2 int64 `json:"character_id2,omitempty"`
 	// ServerRegion holds the value of the "server_region" field.
 	ServerRegion string `json:"server_region,omitempty"`
 	selectValues sql.SelectValues
@@ -63,25 +63,25 @@ func (_m *Bond) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field game_id", values[i])
 			} else if value.Valid {
-				_m.GameID = int(value.Int64)
+				_m.GameID = value.Int64
 			}
 		case bond.FieldGroupID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field group_id", values[i])
 			} else if value.Valid {
-				_m.GroupID = int(value.Int64)
+				_m.GroupID = value.Int64
 			}
 		case bond.FieldCharacterId1:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field character_id1", values[i])
 			} else if value.Valid {
-				_m.CharacterId1 = int(value.Int64)
+				_m.CharacterId1 = value.Int64
 			}
 		case bond.FieldCharacterId2:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field character_id2", values[i])
 			} else if value.Valid {
-				_m.CharacterId2 = int(value.Int64)
+				_m.CharacterId2 = value.Int64
 			}
 		case bond.FieldServerRegion:
 			if value, ok := values[i].(*sql.NullString); !ok {
